feat: 抽离构造YakitTextArea
