+++ app/renderer/src/main/src/pages/YakRunner/RightAuditDetail/RightAuditDetail.module.scss
@@ -1,0 +1,3 @@
+.right-audit-detail{
+    
+}
