--- app/renderer/src/main/src/pages/YakRunner/RightSideBar/RightSideBarType.d.ts
@@ -1,3 +1,0 @@
-export interface RightSideBarProps {}
-
-export type RightSideType = "audit-code" | undefined
